+++ src/main/java/walkingkooka/spreadsheet/reference/SpreadsheetColumnOrRowReferenceRange.java
@@ -62,9 +62,11 @@ abstract class SpreadsheetColumnOrRowReferenceRange<T extends SpreadsheetColumnO
 
     /**
      * Returns the number of columns or rows in this range.
+     *
+     * @return
      */
     @Override
-    public final int count() {
+    public final long count() {
         return this.end()
                 .value()
                 - this.begin()
